feat: [M3-7048] feedback
